+++ toasting-app-api/src/main/kotlin/team/toasting/controller/dto/response/GetMessageCountResponse.kt
@@ -1,4 +1,4 @@
-package team.toasting.controller.dto
+package team.toasting.controller.dto.response
 
 class GetMessageCountResponse(
     val count: Int
