+++ toasting-app-api/src/main/kotlin/team/toasting/controller/dto/response/MemberInfo.kt
@@ -1,4 +1,4 @@
-package team.toasting.controller.dto
+package team.toasting.controller.dto.response
 
 data class MemberInfo(
     val id: Long,
